refactor SectionFilter to FilterForm
