+++ src/components/react-hook-form/index.ts
@@ -3,6 +3,7 @@ export * from './CheckboxFormField'
 export * from './DateFormField'
 export * from './FieldsetFormField'
 export * from './Form'
+export * from './FilterForm'
 export * from './FormSubmitFeedback'
 export * from './InputFormField'
 export * from './NumberFormField'
