+++ src/components/section/index.ts
@@ -15,7 +15,6 @@ export * from './SectionHeader/SectionHeaderSpacer'
 export * from './SectionHeader/SectionHeaderRow'
 export * from './SectionHeader/SectionHeaderGroup'
 export * from './SectionHeader/SectionHeaderTitle'
-export * from './SectionHeader/SectionFilter'
 export * from './SectionHeader/SectionSearch'
 export * from './SectionItem/SectionListItem'
 export * from './SectionItem/SectionLoadingListItem'
